cli/main.py
import argparse

def setup():
    """Setup the environment."""
    print("Setting up the environment...")
    return "Setup complete"

def deploy(file):
    """Deploy the application with the specified file."""
    print(f"Deploying the application with {file}...")
    return f"Deploy complete with {file}"

def status():
    """Check the status of the application."""
    print("Checking the status of the application...")
    return "Status checked"

def create_parser():
    parser = argparse.ArgumentParser(description='A CLI tool with setup, deploy, and status commands.')

    subparsers = parser.add_subparsers(dest='command')
    subparsers.required = True

    parser_setup = subparsers.add_parser('setup', help='Setup the environment.')
    parser_setup.set_defaults(func=setup)

    parser_deploy = subparsers.add_parser('deploy', help='Deploy the application.')
    parser_deploy.add_argument('file', type=str, help='The file to deploy')
    parser_deploy.set_defaults(func=lambda args: deploy(args.file))

    parser_status = subparsers.add_parser('status', help='Check the status of the application.')
    parser_status.set_defaults(func=status)

    return parser

def main():
    parser = create_parser()
    args = parser.parse_args()
    result = args.func(args)
    print(result)

if __name__ == '__main__':
    main()
